Add files via upload
This project is a case study based on how virtual reality (vr) is applied in the field of gaming and other fields such as education, healthcare, etc and many more! Explore out project by going through the review. Thanks to my operating systems teacher who was so supporting during the course of time of developing this project. My sincere gratitude is always there for my teacher!

PRAMODH CODE CRAFTER